Move from writer.common to writer package
