+++ dd-trace-core/src/main/java/datadog/trace/common/writer/RemoteMapper.java
@@ -1,4 +1,4 @@
-package datadog.trace.common.writer.common;
+package datadog.trace.common.writer;
 
 import static java.nio.charset.StandardCharsets.ISO_8859_1;
 
@@ -6,6 +6,7 @@ import datadog.communication.serialization.Mapper;
 import datadog.trace.api.DDTags;
 import datadog.trace.bootstrap.instrumentation.api.Tags;
 import datadog.trace.bootstrap.instrumentation.api.UTF8BytesString;
+import datadog.trace.common.writer.Payload;
 import datadog.trace.core.CoreSpan;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
