+++ dd-trace-core/src/test/groovy/datadog/trace/common/writer/ddagent/TraceMapperV05PayloadTest.groovy
@@ -6,8 +6,8 @@ import datadog.communication.serialization.msgpack.MsgPackWriter
 import datadog.trace.api.DDId
 import datadog.trace.api.DDTags
 import datadog.trace.api.sampling.PrioritySampling
-import datadog.trace.common.writer.common.Payload
-import datadog.trace.common.writer.common.TraceGenerator
+import datadog.trace.common.writer.Payload
+import datadog.trace.common.writer.TraceGenerator
 import datadog.trace.bootstrap.instrumentation.api.Tags
 import datadog.trace.core.DDSpanContext
 import datadog.trace.test.util.DDSpecification
@@ -21,7 +21,7 @@ import java.nio.channels.WritableByteChannel
 import java.util.concurrent.atomic.AtomicInteger
 
 import static datadog.trace.bootstrap.instrumentation.api.InstrumentationTags.DD_MEASURED
-import static datadog.trace.common.writer.common.TraceGenerator.generateRandomTraces
+import static datadog.trace.common.writer.TraceGenerator.generateRandomTraces
 import static org.junit.Assert.assertEquals
 import static org.junit.Assert.assertFalse
 import static org.msgpack.core.MessageFormat.*
